+++ spec/dummy/app/controllers/application_controller.rb
@@ -1,3 +1,5 @@
 class ApplicationController < ActionController::Base
   # Attempt to find the mapped route for devise based on request path
+
+  include AuthSessionConcern
 end