+++ spec/support/omniauth.rb
@@ -1,0 +1,9 @@
+require "omniauth"
+
+OmniAuth.config.test_mode = true
+
+RSpec.configure do |config|
+  config.before do
+    OmniAuth.config.add_mock(:catalogue_patron, JSON.parse(IO.read("spec/files/auth/auth_hash.json")))
+  end
+end